fix compile error from changing int to double type for CourseMark mark

+++ src/main/java/com/sombra/promotion/service/generic/CourseMarkService.java
@@ -33,7 +33,7 @@ public class CourseMarkService {
         CourseMark courseMark = new CourseMark();
         courseMark.setId(uuidUtil.randomUUID());
         courseMark.setStudentId(studentId);
-        courseMark.setMark((int) mark);
+        courseMark.setMark(mark);
         courseMark.setCourseId(courseId);
         return courseMark;
     }
